G_GUINT64_FORMAT expects values of type gint64
Add cast to gint64 to address the following warnings on OS X:

file_io_ops.c: In function ‘fileio_read_on_con_cb’:
file_io_ops.c:751: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 2 has type ‘off_t’
file_io_ops.c:751: warning: format ‘%lu’ expects type ‘long unsigned int’, but argument 3 has type ‘long long unsigned int’

+++ src/file_io_ops.c
@@ -748,7 +748,7 @@ static void fileio_read_on_con_cb (gpointer client, gpointer ctx)
 
         rdata->request_offset = rdata->off;
         range_hdr = g_strdup_printf ("bytes=%"G_GUINT64_FORMAT"-%"G_GUINT64_FORMAT, 
-            rdata->request_offset, rdata->request_offset + part_size);
+            (gint64)rdata->request_offset, (gint64)(rdata->request_offset + part_size));
         http_connection_add_output_header (con, "Range", range_hdr);
         g_free (range_hdr);
     }
